requirements.txt
# MalSentinel — requirements.txt
# Install with: pip install -r requirements.txt

# Core ML
lightgbm>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# PE parsing
pefile>=2023.2.7

# Explainability
shap>=0.44.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Web UI
streamlit>=1.35.0

# PDF reports
fpdf2>=2.7.6

# Hyperparameter tuning (optional but listed in proposal)
optuna>=3.5.0

# Data loading
pyarrow>=14.0.0     # for parquet support

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
networkx>=3.0       # Bayesian network visualisation fallback
pyyaml>=6.0
